Update charmstore tests to avoid state and use gomock too
